Fix conversion warning

+++ solver/src/states/waitingstate.cpp
@@ -28,7 +28,7 @@ void WaitingState::on(SolverContext &context) {
   bSignatures.reserve(counter);
   for (size_t i=0; i< rMask.size(); ++i){ //auto& it : context.final_stage3_data()) {
     if(rMask[i] != cs::ConfidantConsts::InvalidConfidantIndex) {
-      bSignatures.push_back(context.stage3(i)->blockSignature);//emplace_back(it.sender, it.blockSignature);    
+      bSignatures.push_back(context.stage3((uint8_t)i)->blockSignature);//emplace_back(it.sender, it.blockSignature);    
     }
   }
